More Quest Fixes (#102)
* Remove Elite Catalyst Requirement 

Removes Elite Catalyst requirement for Basic Microverse Projector quest, as quantum compressors are no longer needed.

* More Quest Fixes

Changes titanium quest to say kanthal instead of nichrome coils
Updated Primal Mana quest to say Extractor instead of fluid extractor

* Text Error

I added an extra n on line 14818 on accident. Has been removed.

* New Cable Coverings

Allows Styrene-Butadiene Rubber to be used to complete the quest, and adds a description on how to make SBR, similar to the description on how to make Silicon Rubber.

* Oil Quest Changes

Allows all four oil types to be used to complete the navigating petroleum quest. Adds a tooltip giving the difference between the oil types.

* CPU Quest and Bad Formatting

Fixed up some weird formatting caused by the BQ editor, and added the CPU quest, which needs the cleanroom.  Removes the cleanroom as a prerequisite for first tier four circuits, and added the CPU quest prerequisite for The Last Tier One and The Last Tier Two circuits.

* Added reward for CPUs quest

1x Omninickel [5]

* More Formatting Changes

The BQ editor... Also, I forgot to save half of the changes on commit e0893d4...

* HV Machine Hull Quest

Oops. It needed the cleanroom, and the cleanroom needed it...

* Revert all changes relating to the cleanroom.

* Set task logic to OR, turned edit mode off

* Distillation Tower Quest

Changed 'Fluid Input Hatch must go in the center of the bottom layer' to 'Fluid Input Hatch can go anywhere on the bottom layer'