Thermal and AR Stuff, Fix issue #240
Removes all thermal ores, and dyes, from JEI
Adds a lens for the Orbital Laser Drill, and renamed it to avoid confusion. Furthermore, added a warning to the Orbital Laser Drill
Adds the Void World to the AR Database, orbiting around Earth, allowing for rockets launching from the void world to the moon.
Fixes issue #240 by changing the recipes from a fixed ingredient to an oreDict.