Fix Tool Belt's lang via Bansoukou
This was actually a PR (merged) to the Tool Belt repo, but the author of Tool Belt has said that no more updates will b e posted for 1.12.2, unless there is a gamebreaking crash/bug.
Furthermore, improves the README links to the original repos, now pointing them to the branch we are using (eg. 1.12.2)