Pack Mode Switcher changes to address #241 (#292)
Difficulty cannot be inforced by the difficulty locker mod. Meaning it's not only not required by the server, but the difficulty needs to be manually set by server.properties.

I added a second server.properties.expert that just sets difficulty to 0 initially. Obviously a user can change this but they should just know that expert is supposed to be peaceful at that point. I then updated the script to switch between the two configs, if and only if they exist, so it should still work on ssp just the same.

This was made to address #241 Since it is impossible for a mod to enforce server difficulty (Source: Trust me and dig through the code yourself I suggest starting with Forge's DedicatedServer object)

/* Commits: */
* Pack Mode Switcher changes to address 241 in the most roundabout way possible

* pack mode switcher updates and changing server.properties

* added the properties files

* made the readme's more friendly

* more readme-bility

* Made README more tutorial

* typofix/capitialization is my passion

* readme updates

* forced

* repalace

* expert mode, again

* script disclaimer

* Script disclaimer

* hopefully fixed jank formatting

* fixed jank formatting

* beautify