Jetpack QoL
Changes from very slow gradual drop to zero gradual drop with Vibrant and Resonant jetpacks. This is to fix a major pain point in hard mode where no drop flight is not available until t4½mm because of soul binder>angel ring needing heads. No balance impact but major QoL when building, especially in the void.

Removes redundant status display of engine and hover status. Hover is obvious in use and engine on/off is not a functional keybind by default (G but doesn't work due to a conflict).