Add Typo Checking CI Tasks (#1255)
This PR is heavily inspired by the great efforts in #1250, and attempts
to pull together a *somewhat* working system, that is reliable, to
checking for typos.

Like #1250, this system uses https://github.com/crate-ci/typos. This
system runs on the `checks` workflow task.

Checks are limited to lang files (including extracted QB titles and
descriptions), with lang files having the formatting from them stripped
(including substitution areas like %1), and files in the `.github`
folder. Whilst #1250 did contain some typo fixes for code, this is not
included in this PR.

Result:
Hopefully a large reduction in typos, especially for the quest book.

[INTERNAL]