Buff Enriched Naquadah for the EM Separator (#1261)
Another really simple PR. Mostly a balance decision to be discussed. 
Implements one of the fixes from #1260 .

-> Buffs enriched naquadah from `15% + 4%` to `1 & 20% + 5%` (from
+320%(UIV) to +800%(LV) based on tier(and +185%/+860% compared to
Thermal centrifuge))

-> Nerfs normal naquadah from 1 to `50% + 3.75%`

This is a huge buff to the output but is balanced since enriched
naquadah has very little uses, and naquadah is way more useful, being
useful as is, and to make naquadria through naqline.

This is not enough to make the EM separator really useful, but at least
it has ONE niche use, getting easy enriched naquadah for trinium coils.
That being said, it's kinda sad that enriched naq doesn't have more uses
too. It would also give more naquadria than the regular naq recipe if
using the fusion recipe to get it, but since it's MK. III and only in
NM, I don't think it will see any use.

/* GitHub Stuff */
[BALANCING]
[DETAILS]
detail = "Buffs Enriched Naquadah from 15% + 4% to 1 Guaranteed + 20% + 5%"
[DETAILS]

---------

Signed-off-by: Alessian Dormal <alessian@alessian.be>